requirements.txt
numpy==2.3.5
pillow==12.0.0
pypng==0.20220715.0
